refactor(configure): typed bridge params, discriminated credentials
