+++ src/database/classes/ConfigureDb/ConfigureDb.class.ts
@@ -1,8 +1,24 @@
 import type { SalveDatabase } from '../../../specs/SalveDatabase.nitro';
-import type { IConfigureProps, IRegisterProps } from './types';
+import type { IConfigureProps, ICredentialsDefinition, IRegisterProps } from './types';
+import type { IConfigureParams } from '../../../specs/types/IConfigureParams';
 import { NitroModules } from 'react-native-nitro-modules';
 
 const _bridge = NitroModules.createHybridObject<SalveDatabase>('SalveDatabase');
+
+function mapCredentials(creds: ICredentialsDefinition): IConfigureParams['credentials'] {
+  switch (creds.provider) {
+    case 'oauth2':
+      return {
+        provider: 'oauth2',
+        accessTokenHeaderName: creds.accessToken?.headerName ?? 'Authorization',
+        refresh: {
+          endpoint: creds.refresh.endpoint,
+          responseAccessTokenPath: creds.refresh.response.accessToken,
+          responseRefreshTokenPath: creds.refresh.response.refreshToken,
+        },
+      };
+  }
+}
 
 export class ConfigureDb {
   private static _configured = false;
@@ -10,26 +26,33 @@ export class ConfigureDb {
   constructor() {}
 
   configure(props: IConfigureProps): void {
-    const { name } = props;
-    
-    if (!name || name.trim() === '') {
+    if (!props.name || props.name.trim() === '') {
       throw new Error("Database.configure: 'name' is required");
     }
-    _bridge.configure(JSON.stringify(props));
+
+    _bridge.configure({
+      name: props.name,
+      baseUrl: props.baseUrl,
+      network: props.network,
+      credentials: props.credentials !== undefined
+        ? mapCredentials(props.credentials)
+        : undefined,
+    });
+
     ConfigureDb._configured = true;
   }
 
   register(props: IRegisterProps): Promise<void> {
     const { schema } = props;
-    
+
     this._assertConfigured('register');
-    
+
     if (!schema.name || typeof schema.version !== 'number' || !schema.primaryKey) {
       throw new Error(
         "Database.register: schema must have 'name', 'version', and 'primaryKey'"
       );
     }
-    
+
     if (!(schema.primaryKey in schema.columns)) {
       throw new Error(
         `Database.register: primaryKey '${String(schema.primaryKey)}' must exist in schema.columns`
