+++ src/database/classes/ConfigureDb/types/IRegisterProps.ts
@@ -1,4 +1,4 @@
-import type { AnySchema } from "../../../../types/schema/any-schema";
+import type { AnySchema } from "../../../../types";
 
 export interface IRegisterProps {
   schema: AnySchema;
